+++ cmd/compose/c2pcr.yaml
