+++ controllers/composer/testdata/expected/c2pcr-parser-composed-policies/install-odf-lvm-operator.yaml
