+++ controllers/composer/testdata/expected/c2pcr-parser-composed-policies/policy-nginx-deployment.yaml
