+++ controllers/composer/testdata/expected/composed-config-policies/add-chrony/add-chrony.yaml
