+++ controllers/composer/testdata/oscal/component-definition.json
