+++ controllers/composer/testdata/policies/add-chrony/add-chrony-worker/MachineConfig.50-worker-chrony.0.yaml
