+++ controllers/composer/testdata/policies/install-odf-lvm-operator/policy-odf-lvm-operator/OperatorGroup.openshift-storage-operatorgroup.0.yaml
