+++ controllers/composer/testdata/policies/policy-nginx-deployment/kustomization.yaml
