+++ controllers/composer/testdata/policies/policy-nginx-deployment/policy-generator.yaml
