+++ controllers/composer/testdata/policies/policy-nginx-deployment/policy-nginx-deployment/Deployment.nginx-deployment.0.yaml
