+++ pkg/composer/composer.go
@@ -18,194 +18,150 @@ package composer
 
 import (
 	"fmt"
-	"os"
 	"strings"
 
-	policygenerator "github.com/IBM/compliance-to-policy/pkg/policygenerator"
-	. "github.com/IBM/compliance-to-policy/pkg/types/internalcompliance"
-	pgtype "github.com/IBM/compliance-to-policy/pkg/types/policygenerator"
-	cp "github.com/otiai10/copy"
-	typekustomize "sigs.k8s.io/kustomize/api/types"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/sets"
 
 	"github.com/IBM/compliance-to-policy/pkg"
+	"github.com/IBM/compliance-to-policy/pkg/oscal"
+	policygenerator "github.com/IBM/compliance-to-policy/pkg/policygenerator"
+	typec2pcr "github.com/IBM/compliance-to-policy/pkg/types/c2pcr"
+	pgtype "github.com/IBM/compliance-to-policy/pkg/types/policygenerator"
+	cp "github.com/otiai10/copy"
 	"go.uber.org/zap"
-	"k8s.io/apimachinery/pkg/util/sets"
+	"sigs.k8s.io/kustomize/api/resmap"
+	typekustomize "sigs.k8s.io/kustomize/api/types"
+	"sigs.k8s.io/kustomize/kyaml/resid"
 )
 
 var logger *zap.Logger = pkg.GetLogger("composer")
 
-type Composer struct {
+type ComposerV2 struct {
 	policiesDir string
 	tempDir     pkg.TempDirectory
 }
 
-func NewComposer(policiesDir string, tempDir string) *Composer {
-	return NewComposerByTempDirectory(policiesDir, pkg.NewTempDirectory(tempDir))
+func NewComposerV2(policiesDir string, tempDir string) *ComposerV2 {
+	return NewComposerV2ByTempDirectory(policiesDir, pkg.NewTempDirectory(tempDir))
 }
 
-func NewComposerByTempDirectory(policiesDir string, tempDir pkg.TempDirectory) *Composer {
-	return &Composer{
+func NewComposerV2ByTempDirectory(policiesDir string, tempDir pkg.TempDirectory) *ComposerV2 {
+	return &ComposerV2{
 		policiesDir: policiesDir,
 		tempDir:     tempDir,
 	}
 }
 
-func (c *Composer) GetPoliciesDir() string {
+func (c *ComposerV2) GetPoliciesDir() string {
 	return c.policiesDir
 }
 
-type ControlDirectory struct {
-	Path      string
-	ControlId string
-	Policies  []string
+func (c *ComposerV2) ComposeByC2PParsed(c2pParsed typec2pcr.C2PCRParsed) error {
+	return c.Compose(c2pParsed.Namespace, c2pParsed.ComponentObjects, c2pParsed.ClusterSelectors)
 }
 
-type ControlPolicy struct {
-	Path                    string
-	PolicyId                string
-	ConfigPolicies          []string
-	Kustomization           File
-	PolicyGeneratorManifest File
-}
-
-type ControlConfigPolicy struct {
-	Path       string
-	SourcePath string
-}
-type File struct {
-	Path       string
-	SourcePath string
-}
-
-func (c *Composer) Compose(namespace string, compliance Compliance, clusterSelectors map[string]string) (*ComposedResult, error) {
+func (c *ComposerV2) Compose(namespace string, componentObjects []oscal.ComponentObject, clusterSelectors map[string]string) error {
 
 	if clusterSelectors == nil {
 		clusterSelectors = map[string]string{"env": "dev"}
 	}
-	policyCompositions := []PolicyComposition{}
 
-	result := ComposedResult{}
+	logger.Info("Start composing policySets")
+	parameters := map[string]string{}
+	policyConfigMap := map[string]pgtype.PolicyConfig{}
+	policySets := []pgtype.PolicySetConfig{}
+	policySetPatches := []typekustomize.Patch{}
+	for _, componentObject := range componentObjects {
+		logger := logger.With(zap.Namespace(fmt.Sprintf("component %s", componentObject.ComponentTitle)))
+		logger.Info("Start generating policy")
+		for _, ruleObject := range componentObject.RuleObjects {
+			sourceDir := fmt.Sprintf("%s/%s", c.policiesDir, ruleObject.PolicyId)
+			destDir := fmt.Sprintf("%s/%s", c.tempDir.GetTempDir(), ruleObject.PolicyId)
+			err := cp.Copy(sourceDir, destDir)
+			if err != nil {
+				return err
+			}
+		}
 
-	count := 0
-	standard := compliance.Standard
-	for _, category := range standard.Categories {
-		for _, control := range category.Controls {
-			for _, policy := range control.ControlRefs {
-				logger.Info(fmt.Sprintf("Start generating policy '%s'", policy))
-
-				sourceDir := fmt.Sprintf("%s/%s", c.policiesDir, policy)
-				policyCompositionDir := fmt.Sprintf("%s/%s", c.tempDir.GetTempDir(), policy)
-				err := cp.Copy(sourceDir, policyCompositionDir)
-				if err != nil {
-					return nil, err
-				}
-				policyGeneratorManifestPath := policyCompositionDir + "/policy-generator.yaml"
-				var policyGeneratorManifest pgtype.PolicyGenerator
-				if err := pkg.LoadYamlFileToObject(sourceDir+"/policy-generator.yaml", &policyGeneratorManifest); err != nil {
-					return nil, err
-				}
-				policyGeneratorManifest.PolicyDefaults.Namespace = namespace
-				policyGeneratorManifest.PolicyDefaults.PolicyOptions.Standards = []string{standard.Name}
-				policyGeneratorManifest.PolicyDefaults.PolicyOptions.Categories = []string{category.Name}
-				policyGeneratorManifest.PolicyDefaults.PolicyOptions.Controls = []string{control.Name}
-				policyGeneratorManifest.PolicyDefaults.PolicyOptions.Placement.ClusterSelectors = clusterSelectors
-				if err := pkg.WriteObjToYamlFileByGoYaml(policyGeneratorManifestPath, policyGeneratorManifest); err != nil {
-					return nil, err
-				}
-				logger.Info(fmt.Sprintf("Generate policy '%s' by PolicyGenerator", policyGeneratorManifestPath))
-				generatedManifests, err := policygenerator.Kustomize(policyCompositionDir)
-				if err != nil {
-					logger.Sugar().Error(err, "failed to run kustomize")
-					return nil, err
-				}
-				entries, err := os.ReadDir(policyCompositionDir)
-				if err != nil {
-					return nil, err
-				}
-				configPolicyAbsoluteDirs := []string{}
-				for _, entry := range entries {
-					if entry.IsDir() {
-						configPolicyAbsoluteDirs = append(configPolicyAbsoluteDirs, policyCompositionDir+"/"+entry.Name())
+		for idx, controlImpleObject := range componentObject.ControlImpleObjects {
+			policyListPerControlImple := []string{}
+			for _, param := range controlImpleObject.SetParameters {
+				parameters[param.ParamID] = param.Values[0]
+			}
+			for _, controlObject := range controlImpleObject.ControlObjects {
+				for _, ruleId := range controlObject.RuleIds {
+					ruleObject, ok := oscal.FindRulesByRuleId(ruleId, componentObject.RuleObjects)
+					if ok {
+						policyId := ruleObject.PolicyId
+						destDir := fmt.Sprintf("%s/%s", c.tempDir.GetTempDir(), policyId)
+						policyGeneratorManifestPath := destDir + "/policy-generator.yaml"
+						var policyGeneratorManifest pgtype.PolicyGenerator
+						if err := pkg.LoadYamlFileToObject(policyGeneratorManifestPath, &policyGeneratorManifest); err != nil {
+							return err
+						}
+						policyGeneratorManifest.PolicyDefaults.Namespace = namespace
+						policyGeneratorManifest.PolicyDefaults.PolicyOptions.Standards = []string{""}
+						policyGeneratorManifest.PolicyDefaults.PolicyOptions.Categories = []string{""}
+						policyGeneratorManifest.PolicyDefaults.PolicyOptions.Controls = []string{controlObject.ControlId}
+						policyGeneratorManifest.PolicyDefaults.PolicyOptions.Placement.ClusterSelectors = clusterSelectors
+						if err := pkg.WriteObjToYamlFileByGoYaml(policyGeneratorManifestPath, policyGeneratorManifest); err != nil {
+							return err
+						}
+						// For policySet
+						policyListPerControlImple = appendUnique(policyListPerControlImple, policyId)
+						policyConfig, ok := policyConfigMap[policyId]
+						if ok {
+							policyConfig.Standards = appendUnique(policyConfig.Standards, policyGeneratorManifest.PolicyDefaults.Standards...)
+							policyConfig.Categories = appendUnique(policyConfig.Categories, policyGeneratorManifest.PolicyDefaults.Categories...)
+							policyConfig.Controls = appendUnique(policyConfig.Controls, policyGeneratorManifest.PolicyDefaults.Controls...)
+							policyConfigMap[policyId] = policyConfig
+						} else {
+							policyConfig := policyGeneratorManifest.Policies[0]
+							policyConfig.Standards = policyGeneratorManifest.PolicyDefaults.Standards
+							policyConfig.Categories = policyGeneratorManifest.PolicyDefaults.Categories
+							policyConfig.Controls = policyGeneratorManifest.PolicyDefaults.Controls
+							for idx, manifest := range policyConfig.Manifests {
+								policyConfig.Manifests[idx].Path = strings.Replace(manifest.Path, "./", fmt.Sprintf("./%s/", policyId), 1)
+							}
+							policyConfigMap[policyId] = policyConfig
+						}
 					}
 				}
-				policyComposition := PolicyComposition{
-					Id:                      policy,
-					ControlId:               control.Name,
-					PolicyCompositionDir:    policyCompositionDir,
-					configPolicyDirs:        configPolicyAbsoluteDirs,
-					composedManifests:       &generatedManifests,
-					policyGeneratorManifest: policyGeneratorManifest,
-				}
-				policyCompositions = append(policyCompositions, policyComposition)
-				logger.Info(fmt.Sprintf("Finish generating policy '%s'", policy))
-				count = count + 1
 			}
+			suffix := ""
+			if idx > 0 {
+				suffix = fmt.Sprintf("-%d", idx)
+			}
+			policySetConfig := pgtype.PolicySetConfig{
+				Name:     toDNSCompliant(componentObject.ComponentTitle + suffix),
+				Policies: policyListPerControlImple,
+			}
+			policySets = append(policySets, policySetConfig)
+			policySetPatch := typekustomize.Patch{
+				Target: &typekustomize.Selector{
+					ResId: resid.FromString(fmt.Sprintf("PolicySet../%s.", policySetConfig.Name)),
+				},
+				Patch: fmt.Sprintf(`[{"op": "replace", "path": "/metadata/annotations/%s", "value": "%s"}]`, pkg.ANNOTATION_COMPONENT_TITLE, componentObject.ComponentTitle),
+			}
+			policySetPatches = append(policySetPatches, policySetPatch)
 		}
 	}
-	result.policyCompositions = policyCompositions
-	result.internalCompliance = compliance
-	result.namespace = namespace
-	result.clusterSelectors = clusterSelectors
 
-	policySetsGeneratorManifest := generatePolicySetsGeneratorManifest(&result)
-	if err := pkg.WriteObjToYamlFileByGoYaml(c.tempDir.GetTempDir()+"/policy-generator.yaml", policySetsGeneratorManifest); err != nil {
-		return nil, err
-	}
-	kustomize := typekustomize.Kustomization{Generators: []string{"./policy-generator.yaml"}}
-	if err := pkg.WriteObjToYamlFile(c.tempDir.GetTempDir()+"/kustomization.yaml", kustomize); err != nil {
-		return nil, err
-	}
-
-	generatedManifests, err := policygenerator.Kustomize(c.tempDir.GetTempDir())
-	if err != nil {
-		logger.Sugar().Error(err, "failed to run kustomize")
-		return nil, err
-	}
-	result.composedManifests = &generatedManifests
-
-	logger.Info("")
-	logger.Info(fmt.Sprintf("%d policies are created", count))
-
-	return &result, nil
-}
-
-func generatePolicySetsGeneratorManifest(cr *ComposedResult) pgtype.PolicyGenerator {
 	policyDefaults := pgtype.PolicyDefaults{
-		Namespace: cr.namespace,
+		Namespace: namespace,
 		PolicyOptions: pgtype.PolicyOptions{
 			Placement: pgtype.PlacementConfig{
-				ClusterSelectors: cr.clusterSelectors,
+				LabelSelector: clusterSelectors,
 			},
 		},
-	}
-	policyConfigMap := map[string]pgtype.PolicyConfig{}
-	policyListPerControl := map[string][]string{}
-	for _, policyComposition := range cr.policyCompositions {
-		policyName := policyComposition.Id
-		controlId := policyComposition.ControlId
-		policyList, ok := policyListPerControl[controlId]
-		if ok {
-			policyListPerControl[controlId] = append(policyList, policyName)
-		} else {
-			policyListPerControl[controlId] = []string{policyName}
-		}
-		policyConfig, ok := policyConfigMap[policyName]
-		policyGeneratorManifest := policyComposition.policyGeneratorManifest
-		if ok {
-			policyConfig.Standards = appendUnique(policyConfig.Standards, policyGeneratorManifest.PolicyDefaults.Standards...)
-			policyConfig.Categories = appendUnique(policyConfig.Categories, policyGeneratorManifest.PolicyDefaults.Categories...)
-			policyConfig.Controls = appendUnique(policyConfig.Controls, policyGeneratorManifest.PolicyDefaults.Controls...)
-			policyConfigMap[policyName] = policyConfig
-		} else {
-			policyConfig := policyGeneratorManifest.Policies[0]
-			policyConfig.Standards = policyGeneratorManifest.PolicyDefaults.Standards
-			policyConfig.Categories = policyGeneratorManifest.PolicyDefaults.Categories
-			policyConfig.Controls = policyGeneratorManifest.PolicyDefaults.Controls
-			for idx, manifest := range policyConfig.Manifests {
-				policyConfig.Manifests[idx].Path = strings.Replace(manifest.Path, "./", fmt.Sprintf("./%s/", policyName), 1)
-			}
-			policyConfigMap[policyName] = policyConfig
-		}
+		ConfigurationPolicyOptions: pgtype.ConfigurationPolicyOptions{
+			NamespaceSelector: pgtype.NamespaceSelector{
+				Exclude: []string{"kube-system", "open-cluster-management", "open-cluster-management-agent", "open-cluster-management-agent-addon"},
+				Include: []string{"*"},
+			},
+		},
 	}
 	policyConfigs := []pgtype.PolicyConfig{}
 	for _, policyConfig := range policyConfigMap {
@@ -213,23 +169,47 @@ func generatePolicySetsGeneratorManifest(cr *ComposedResult) pgtype.PolicyGenera
 	}
 	policySetGeneratorManifest := policygenerator.BuildPolicyGeneratorManifest("policy-set", policyDefaults, policyConfigs)
 	policySetGeneratorManifest.PlacementBindingDefaults.Name = "policy-set"
-	policySetGeneratorManifest.PolicySets = []pgtype.PolicySetConfig{}
-	for controlId, policyList := range policyListPerControl {
-		policySetConfig := pgtype.PolicySetConfig{
-			Name:     controlId,
-			Policies: policyList,
-		}
-		policySetGeneratorManifest.PolicySets = append(policySetGeneratorManifest.PolicySets, policySetConfig)
+	policySetGeneratorManifest.PolicySets = policySets
+	policySetGeneratorManifest.PolicySetDefaults = pgtype.PolicySetDefaults{
+		PolicySetOptions: pgtype.PolicySetOptions{
+			Placement: policyDefaults.Placement,
+		},
 	}
-	return policySetGeneratorManifest
+
+	if err := pkg.WriteObjToYamlFileByGoYaml(c.tempDir.GetTempDir()+"/policy-generator.yaml", policySetGeneratorManifest); err != nil {
+		return err
+	}
+
+	logger.Info("Create configmapt for templatized parameters")
+	parametersConfigmap := corev1.ConfigMap{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "ConfigMap",
+			APIVersion: "v1",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "c2p-parameters",
+			Namespace: "c2p",
+		},
+		Data: parameters,
+	}
+	if err := pkg.WriteObjToYamlFile(c.tempDir.GetTempDir()+"/parameters.yaml", parametersConfigmap); err != nil {
+		return err
+	}
+
+	kustomize := typekustomize.Kustomization{
+		Generators: []string{"./policy-generator.yaml"},
+		Resources:  []string{"./parameters.yaml"},
+		Patches:    policySetPatches,
+	}
+	if err := pkg.WriteObjToYamlFile(c.tempDir.GetTempDir()+"/kustomization.yaml", kustomize); err != nil {
+		return err
+	}
+	logger.Info("")
+
+	return nil
 }
 
-func appendUnique(slice []string, elems ...string) []string {
-	a := append(slice, elems...)
-	return sets.List[string](sets.New[string](a...))
-}
-
-func (c *Composer) CopyAllTo(destDir string) error {
+func (c *ComposerV2) CopyAllTo(destDir string) error {
 	if _, err := pkg.MakeDir(destDir); err != nil {
 		return err
 	}
@@ -237,4 +217,25 @@ func (c *Composer) CopyAllTo(destDir string) error {
 		return err
 	}
 	return nil
+}
+
+func (c *ComposerV2) GeneratePolicySet() (*resmap.ResMap, error) {
+	generatedManifests, err := policygenerator.Kustomize(c.tempDir.GetTempDir())
+	if err != nil {
+		logger.Sugar().Error(err, "failed to run kustomize")
+		return nil, err
+	}
+	return &generatedManifests, nil
+}
+
+func toDNSCompliant(name string) string {
+	var result string
+	result = strings.ToLower(name)
+	result = strings.ReplaceAll(result, " ", "-")
+	return result
+}
+
+func appendUnique(slice []string, elems ...string) []string {
+	a := append(slice, elems...)
+	return sets.List[string](sets.New[string](a...))
 }
