--- pkg/composer/testdata/compliance.yaml
@@ -1,9 +1,0 @@
-standard:
-  name: NIST SP 800-53
-  categories:
-  - name: CM Configuration Management
-    controls:
-    - name: CM-2 Baseline Configuration
-      controlrefs:
-      - add-chrony
-      - install-odf-lvm-operator
